Add metrics system to measure usage and success of the library
The signaling backend will add a metrics interface to report events to
the configured METRICS_URL and records lobby events and some rtc events.
The clients passes most other rtc events over the websocket to also be
recorded.

The receiving end of the metrics is left open, and the format is to be
documented (read the code for now...).