Fix reconnect loop when reconnecting with bad secret
Right now clients get stuck in a reconnect loop when their secret is
bad. This change makes it so they don't immediately reconnect. WebRTC
connections will stay in tact, so players will be able to keep playing.
But they won't be able to list or join other lobbies without reloading
the game.