Improve CreateLobby, JoinLobby and LeaveLobby
Improve these functions by adding a subscription on the lobby itself
instead of only the lobby in combination with the peer. This makes it so
these methods don't have to return lists of peers anymore.

Also fix the TODO to add a peer to a lobby in CreateLobby instead of
calling a separate JoinLobby afterwards.

Also fixes the broken reconnect code which didn't support multiple
lobbies yet.