Make createPlayer have a small delay (#98)
Give the Network a little bit of time to connect to the signalling
backend before we continue with the next step. This makes our tests less
flaky. Found out when some lobby tests where we create multiple players
after each other was failing because their ID was changing (as they are
given out in order).