Fix test data empty array and NULL value inserts (#99)
Inserting `[]` into postgress will give an error that it doesn't know
the type of the array. Since we're always inserting arrays of strings
here we can specify the type here.

Also converts `null` to insert `NULL` instead of the string `'NULL'`.

These are currently no in use, and because of that changes to the tests
aren't needed. But I want to introduce tests with these in the Lobby
Leader pull, and want to keep that pull simple and small.