Fix reconnects not being recognized (#100)
If we reconnect too quickly, the signalling server won't have the time
to add our information to the database yet and the reconnect won't
recognize us. This can be fixed by waiting a little bit before we
reconnect.

When we only disconnect for shortly it will take the signaling server
longer to notice the disconnect than to reconnect. In which case the
reconnect isn't recognized as a disconnect.

This is still not a perfect solution and best-effort only. But from my
testing where I noticed this happening it was fixed this way for now.