Leader election (#93)
Give each lobby a leader. A leader is a useful concept to have in many
game types. Implementing leader election in the game itself is much
harder.

After this we can add support for the leader to be able to change the
lobby data. For example making it private when a game starts.