Detect inactive peers and remove them from lobbies
Inactive peers are peers that haven't had a connection to the signalling
server for a while.

Normally the TimeoutManager keeps track of which peers have disconnected
and removes the ones that don't reconnect within a certain time.
But this tracking sometimes fails which leaves us with peers in lobbies
that are never coming back.

This pull adds functionality to the TimeoutManager to track activity of
all peers and make sure they are always removed if they aren't active
for a while.

This doesn't mean we can remove the old functionality of the
TimeoutManager. This is still used to allow peers to reconnect using the
same peer ID.