Ping the signaling server from the client
Send pings to be able to detect broken connections. No need for a pong,
TCP will take care of that.

Log the origin with "unknown packet type received" so we can see which
clients are sending packets the server doesn't support (in this case it
was ping from old clients?).