Remove peer MaxConnectionTime (#190)
There is no need to always disconnect the Websocket connection after an
hour. This just causes issues once in a while with db queries being
executed at the same time and then being cancelled.

**Edit:** it's actually causing issues quite often
with`MarkPeerAsActive`. This function is scheduled every 30 seconds with
a ticker. Because it's a ticker and keeps rhythm perfectly it is always
running right after the hour and almost always causes a failed with
`context deadline exceeded` error.