Prevent sql injection from string concatenation (#217)
When a `description` packet is received receipient is passed as argument
to Publish. This field is not validated and could contain SQL
injections. We could validate the field, but a vunerability like this
could happen again in the future, therefor it's better to validate the
topic in Publish itself.

See:
https://github.com/poki/netlib/blob/0dbfcd849e81458205d0bf198e3b03d2cee501ba/internal/signaling/peer.go#L166

This issue was found by OpenAI o3.