Another latency experiment (#299)
The previous latency calculations are still being evaluated, but I would
like to also test latency latency calculations based on lat/long
coordinates.

In this experimenal code we'll get the lat/long from cloudflare and use
that to estimate the latency between peers.

We can then evaluate this value next to the previous latency value based
on base station pings, and the latency we measure between peers.

We'll also store the lat/long coordinates with the `avg-latency-at-10s`
event. This way we can start building a map of latencies between
coordinates in the world. We could use this map to more accurately
predict latencies in the future.