Store country and region with 10s latency events (#312)
Lat,long seems to result in a lot of bad data. I now want to try and use
countries and regions to predict latencies instead. We don't want to
store the region for each peer, in small countries such as the
netherlands this is not really needed or relevant. Only for big
countries regions are relevant and (hopefully) accurate.